getting to know things
