+++ game/lobby.go
@@ -1,26 +1,18 @@
 package game
 
 import (
-	"encoding/json"
 	"fmt"
-	"html"
 	"log"
-	"math"
 	"math/rand"
-	"strconv"
-	"strings"
 	"sync"
 	"time"
 
-	commands "github.com/Bios-Marcel/cmdp"
-	"github.com/Bios-Marcel/discordemojimap"
-	"github.com/agnivade/levenshtein"
-	petname "github.com/dustinkirkland/golang-petname"
+	uuid "github.com/satori/go.uuid"
 )
 
 var (
-	createDeleteMutex          = &sync.Mutex{}
-	lobbies           []*Lobby = nil
+	lobbies   []*Lobby
+	lobbiesMu = &sync.Mutex{}
 )
 
 var (
@@ -36,9 +28,180 @@ var (
 	}
 	SupportedLanguages = map[string]string{
 		"english": "English",
-		"french": "French",
+		"french":  "French",
 	}
 )
+
+// Lobby represents a game session.
+type Lobby struct {
+	// ID uniquely identified the Lobby.
+	ID string
+
+	// DrawingTime is the amount of seconds that each player has available to
+	// finish their drawing.
+	DrawingTime int
+	// MaxRounds defines how many iterations a lobby does before the game ends.
+	// One iteration means every participant does one drawing.
+	MaxRounds int
+	// MaxPlayers defines the maximum amount of players in a single lobby.
+	MaxPlayers int
+	// CustomWords are additional words that will be used in addition to the
+	// predefined words.
+	CustomWords []string
+	Words       []string
+
+	// Players references all participants of the Lobby.
+	Players []*Player
+
+	// Drawer references the Player that is currently drawing.
+	Drawer *Player
+	// Owner references the Player that created the lobby.
+	Owner *Player
+	// CurrentWord represents the word that was last selected. If no word has
+	// been selected yet or the round is already over, this should be empty.
+	CurrentWord string
+	// WordHints for the current word.
+	WordHints []*WordHint
+	// WordHintsShown are the same as WordHints with characters visible.
+	WordHintsShown []*WordHint
+	// Round is the round that the Lobby is currently in. This is a number
+	// between 0 and MaxRounds. 0 indicates that it hasn't started yet.
+	Round int
+	// WordChoice represents the current choice of words.
+	WordChoice []string
+	// RoundEndTime represents the time at which the current round will end.
+	// This is a UTC unix-timestamp in milliseconds.
+	RoundEndTime int64
+
+	timeLeftTicker        *time.Ticker
+	timeLeftTickerReset   chan struct{}
+	scoreEarnedByGuessers int
+	alreadyUsedWords      []string
+	CustomWordsChance     int
+	ClientsPerIPLimit     int
+	// CurrentDrawing represents the state of the current canvas. The elements
+	// consist of LineEvent and FillEvent. Please do not modify the contents
+	// of this array an only move AppendLine and AppendFill on the respective
+	// lobby object.
+	CurrentDrawing []interface{}
+	EnableVotekick bool
+}
+
+// WordHint describes a character of the word that is to be guessed, whether
+// the character should be shown and whether it should be underlined on the
+// UI.
+type WordHint struct {
+	Character rune `json:"character"`
+	Underline bool `json:"underline"`
+}
+
+// Line is the struct that a client send when drawing
+type Line struct {
+	FromX     float32 `json:"fromX"`
+	FromY     float32 `json:"fromY"`
+	ToX       float32 `json:"toX"`
+	ToY       float32 `json:"toY"`
+	Color     string  `json:"color"`
+	LineWidth float32 `json:"lineWidth"`
+}
+
+// Fill represents the usage of the fill bucket.
+type Fill struct {
+	X     float32 `json:"x"`
+	Y     float32 `json:"y"`
+	Color string  `json:"color"`
+}
+
+// GetLobby returns a Lobby that has a matching ID or no Lobby if none could
+// be found.
+func GetLobby(id string) *Lobby {
+	for _, l := range lobbies {
+		if l.ID == id {
+			return l
+		}
+	}
+
+	return nil
+}
+
+// RemoveLobby deletes a lobby, not allowing anyone to connect to it again.
+func RemoveLobby(id string) {
+	indexToDelete := -1
+	for index, l := range lobbies {
+		if l.ID == id {
+			indexToDelete = index
+		}
+	}
+
+	if indexToDelete != -1 {
+		lobbies = append(lobbies[:indexToDelete], lobbies[indexToDelete+1:]...)
+	}
+}
+func createLobby(
+	drawingTime int,
+	rounds int,
+	maxPlayers int,
+	customWords []string,
+	customWordsChance int,
+	clientsPerIPLimit int,
+	enableVotekick bool) *Lobby {
+
+	lobby := &Lobby{
+		ID:                  uuid.NewV4().String(),
+		DrawingTime:         drawingTime,
+		MaxRounds:           rounds,
+		MaxPlayers:          maxPlayers,
+		CustomWords:         customWords,
+		CustomWordsChance:   customWordsChance,
+		timeLeftTickerReset: make(chan struct{}),
+		ClientsPerIPLimit:   clientsPerIPLimit,
+		EnableVotekick:      enableVotekick,
+		CurrentDrawing:      make([]interface{}, 0, 0),
+	}
+
+	if len(customWords) > 1 {
+		rand.Shuffle(len(lobby.CustomWords), func(i, j int) {
+			lobby.CustomWords[i], lobby.CustomWords[j] = lobby.CustomWords[j], lobby.CustomWords[i]
+		})
+	}
+
+	lobbiesMu.Lock()
+	defer lobbiesMu.Unlock()
+	lobbies = append(lobbies, lobby)
+
+	return lobby
+}
+
+func (lobby *Lobby) HasConnectedPlayers() bool {
+	for _, p := range lobby.Players {
+		if p.Connected {
+			return true
+		}
+	}
+
+	return false
+}
+
+// CreateLobby allows creating a lobby, optionally returning errors that
+// occured during creation.
+func CreateLobby(playerName, language string, drawingTime, rounds, maxPlayers, customWordChance, clientsPerIPLimit int, customWords []string, enableVotekick bool) (string, *Lobby, error) {
+	lobby := createLobby(drawingTime, rounds, maxPlayers, customWords, customWordChance, clientsPerIPLimit, enableVotekick)
+	player := createPlayer(playerName)
+
+	lobby.Players = append(lobby.Players, player)
+	lobby.Owner = player
+
+	// Read wordlist according to the chosen language
+	words, err := readWordList(language)
+	if err != nil {
+		//TODO Remove lobby, since we errored.
+		return "", nil, err
+	}
+
+	lobby.Words = words
+
+	return player.userSession, lobby, nil
+}
 
 // SettingBounds defines the lower and upper bounds for the user-specified
 // lobby creation input.
@@ -53,7 +216,7 @@ type SettingBounds struct {
 	MaxClientsPerIPLimit int64
 }
 
-// LineEvent is basically the same as JSEvent, but with a specific Data type.
+// LineEvent is basically the same as GameEvent, but with a specific Data type.
 // We use this for reparsing as soon as we know that the type is right. It's
 // a bit unperformant, but will do for now.
 type LineEvent struct {
@@ -61,7 +224,7 @@ type LineEvent struct {
 	Data *Line  `json:"data"`
 }
 
-// LineEvent is basically the same as JSEvent, but with a specific Data type.
+// LineEvent is basically the same as GameEvent, but with a specific Data type.
 // We use this for reparsing as soon as we know that the type is right. It's
 // a bit unperformant, but will do for now.
 type FillEvent struct {
@@ -69,327 +232,33 @@ type FillEvent struct {
 	Data *Fill  `json:"data"`
 }
 
-func HandleEvent(raw []byte, received *JSEvent, lobby *Lobby, player *Player) error {
-	if received.Type == "message" {
-		dataAsString, isString := (received.Data).(string)
-		if !isString {
-			return fmt.Errorf("invalid data received: '%s'", received.Data)
-		}
-
-		if strings.HasPrefix(dataAsString, "!") {
-			handleCommand(dataAsString[1:], player, lobby)
-		} else {
-			handleMessage(dataAsString, player, lobby)
-		}
-	} else if received.Type == "line" {
-		if lobby.canDraw(player) {
-			line := &LineEvent{}
-			jsonError := json.Unmarshal(raw, line)
-			if jsonError != nil {
-				return fmt.Errorf("error decoding data: %s", jsonError)
-			}
-			lobby.AppendLine(line)
-
-			//We directly forward the event, as it seems to be valid.
-			SendDataToConnectedPlayers(player, lobby, received)
-		}
-	} else if received.Type == "fill" {
-		if lobby.canDraw(player) {
-			fill := &FillEvent{}
-			jsonError := json.Unmarshal(raw, fill)
-			if jsonError != nil {
-				return fmt.Errorf("error decoding data: %s", jsonError)
-			}
-			lobby.AppendFill(fill)
-
-			//We directly forward the event, as it seems to be valid.
-			SendDataToConnectedPlayers(player, lobby, received)
-		}
-	} else if received.Type == "clear-drawing-board" {
-		if lobby.canDraw(player) {
-			lobby.ClearDrawing()
-			SendDataToConnectedPlayers(player, lobby, received)
-		}
-	} else if received.Type == "choose-word" {
-		chosenIndex, isInt := (received.Data).(int)
-		if !isInt {
-			asFloat, isFloat32 := (received.Data).(float64)
-			if isFloat32 && asFloat < 4 {
-				chosenIndex = int(asFloat)
-			} else {
-				fmt.Println("Invalid data")
-				return nil
-			}
-		}
-
-		drawer := lobby.Drawer
-		if player == drawer && len(lobby.WordChoice) > 0 && chosenIndex >= 0 && chosenIndex <= 2 {
-			lobby.CurrentWord = lobby.WordChoice[chosenIndex]
-			lobby.WordChoice = nil
-			lobby.WordHints = createWordHintFor(lobby.CurrentWord, false)
-			lobby.WordHintsShown = createWordHintFor(lobby.CurrentWord, true)
-			triggerWordHintUpdate(lobby)
-		}
-	} else if received.Type == "kick-vote" {
-		toKickID, isString := (received.Data).(string)
-		if !isString {
-			fmt.Println("Invalid data")
-			return nil
-		}
-		if !lobby.EnableVotekick {
-			// Votekicking is disabled in the lobby
-			// We tell the user and do not continue with the event
-			WriteAsJSON(player, JSEvent{Type: "system-message", Data: "Votekick is disabled in this lobby!"})
-		} else {
-			handleKickEvent(lobby, player, toKickID)
-		}
-	} else if received.Type == "start" {
-		if lobby.Round == 0 && player == lobby.Owner {
-			for _, otherPlayer := range lobby.Players {
-				otherPlayer.Score = 0
-				otherPlayer.LastScore = 0
-			}
-
-			lobby.Round = 1
-
-			advanceLobby(lobby)
+// GetPlayer searches for a player, identifying them by usersession.
+func (lobby *Lobby) GetPlayer(userSession string) *Player {
+	for _, player := range lobby.Players {
+		if player.userSession == userSession {
+			return player
 		}
 	}
 
 	return nil
 }
 
-func handleMessage(input string, sender *Player, lobby *Lobby) {
-	trimmed := strings.TrimSpace(input)
-	if trimmed == "" {
-		return
-	}
-
-	if lobby.CurrentWord == "" {
-		sendMessageToAll(trimmed, sender, lobby)
-		return
-	}
-
-	if sender.State == Drawing || sender.State == Standby {
-		sendMessageToAllNonGuessing(trimmed, sender, lobby)
-	} else if sender.State == Guessing {
-		lowerCasedInput := strings.ToLower(trimmed)
-		lowerCasedSearched := strings.ToLower(lobby.CurrentWord)
-		if lowerCasedSearched == lowerCasedInput {
-			secondsLeft := lobby.RoundEndTime/1000 - time.Now().UTC().UnixNano()/1000000000
-			sender.LastScore = int(math.Ceil(math.Pow(math.Max(float64(secondsLeft), 1), 1.3) * 2))
-			sender.Score += sender.LastScore
-			lobby.scoreEarnedByGuessers += sender.LastScore
-			sender.State = Standby
-			WriteAsJSON(sender, JSEvent{Type: "system-message", Data: "You have correctly guessed the word."})
-
-			if !lobby.isAnyoneStillGuessing() {
-				endRound(lobby)
-			} else {
-				//Since the word has been guessed correctly, we reveal it.
-				WriteAsJSON(sender, JSEvent{Type: "update-wordhint", Data: lobby.WordHintsShown})
-				recalculateRanks(lobby)
-				triggerCorrectGuessEvent(lobby)
-				triggerPlayersUpdate(lobby)
-			}
-
-			return
-		} else if levenshtein.ComputeDistance(lowerCasedInput, lowerCasedSearched) == 1 {
-			WriteAsJSON(sender, JSEvent{Type: "system-message", Data: fmt.Sprintf("'%s' is very close.", trimmed)})
-		}
-
-		sendMessageToAll(trimmed, sender, lobby)
-	}
+func (lobby *Lobby) ClearDrawing() {
+	lobby.CurrentDrawing = make([]interface{}, 0, 0)
 }
 
-func (lobby *Lobby) isAnyoneStillGuessing() bool {
-	for _, otherPlayer := range lobby.Players {
-		if otherPlayer.State == Guessing && otherPlayer.Connected {
-			return true
-		}
-	}
-
-	return false
+// AppendLine adds a line direction to the current drawing. This exists in order
+// to prevent adding arbitrary elements to the drawing, as the backing array is
+// an empty interface type.
+func (lobby *Lobby) AppendLine(line *LineEvent) {
+	lobby.CurrentDrawing = append(lobby.CurrentDrawing, line)
 }
 
-func sendMessageToAll(message string, sender *Player, lobby *Lobby) {
-	escaped := html.EscapeString(discordemojimap.Replace(message))
-	for _, target := range lobby.Players {
-		WriteAsJSON(target, JSEvent{Type: "message", Data: Message{
-			Author:  html.EscapeString(sender.Name),
-			Content: escaped,
-		}})
-	}
-}
-
-func sendMessageToAllNonGuessing(message string, sender *Player, lobby *Lobby) {
-	escaped := html.EscapeString(discordemojimap.Replace(message))
-	for _, target := range lobby.Players {
-		if target.State != Guessing {
-			WriteAsJSON(target, JSEvent{Type: "non-guessing-player-message", Data: Message{
-				Author:  html.EscapeString(sender.Name),
-				Content: escaped,
-			}})
-		}
-	}
-}
-
-func handleKickEvent(lobby *Lobby, player *Player, toKickID string) {
-	//Kicking yourself isn't allowed
-	if toKickID == player.ID {
-		return
-	}
-
-	//A player can't vote twice to kick someone
-	if player.votedForKick[toKickID] {
-		return
-	}
-
-	toKick := -1
-	for index, otherPlayer := range lobby.Players {
-		if otherPlayer.ID == toKickID {
-			toKick = index
-			break
-		}
-	}
-
-	//If we haven't found the player, we can't kick him/her.
-	if toKick != -1 {
-		player.votedForKick[toKickID] = true
-		playerToKick := lobby.Players[toKick]
-
-		var voteKickCount int
-		for _, otherPlayer := range lobby.Players {
-			if otherPlayer.votedForKick[toKickID] == true {
-				voteKickCount++
-			}
-		}
-
-		votesNeeded := 1
-		if len(lobby.Players)%2 == 0 {
-			votesNeeded = len(lobby.Players) / 2
-		} else {
-			votesNeeded = (len(lobby.Players) / 2) + 1
-		}
-
-		WritePublicSystemMessage(lobby, fmt.Sprintf("(%d/%d) players voted to kick %s", voteKickCount, votesNeeded, playerToKick.Name))
-
-		if voteKickCount >= votesNeeded {
-			//Since the player is already kicked, we first clean up the kicking information related to that player
-			for _, otherPlayer := range lobby.Players {
-				if otherPlayer.votedForKick[toKickID] == true {
-					delete(player.votedForKick, toKickID)
-					break
-				}
-			}
-
-			WritePublicSystemMessage(lobby, fmt.Sprintf("%s has been kicked from the lobby", playerToKick.Name))
-
-			if lobby.Drawer == playerToKick {
-				WritePublicSystemMessage(lobby, "Since the kicked player has been drawing, none of you will get any points this round.")
-				//Since the drawing person has been kicked, that probably means that he/she was trolling, therefore
-				//we redact everyones last earned score.
-				for _, otherPlayer := range lobby.Players {
-					otherPlayer.Score -= otherPlayer.LastScore
-					otherPlayer.LastScore = 0
-				}
-				lobby.scoreEarnedByGuessers = 0
-				//We must absolutely not set lobby.Drawer to nil, since this would cause the drawing order to be ruined.
-			}
-
-			if playerToKick.ws != nil {
-				playerToKick.ws.Close()
-			}
-			lobby.Players = append(lobby.Players[:toKick], lobby.Players[toKick+1:]...)
-
-			recalculateRanks(lobby)
-
-			//If the owner is kicked, we choose the next best person as the owner.
-			if lobby.Owner == playerToKick {
-				for _, otherPlayer := range lobby.Players {
-					potentialOwner := otherPlayer
-					if potentialOwner.Connected {
-						lobby.Owner = potentialOwner
-						WritePublicSystemMessage(lobby, fmt.Sprintf("%s is the new lobby owner.", potentialOwner.Name))
-						break
-					}
-				}
-			}
-
-			triggerPlayersUpdate(lobby)
-
-			if lobby.Drawer == playerToKick || !lobby.isAnyoneStillGuessing() {
-				endRound(lobby)
-			}
-		}
-	}
-}
-
-func handleCommand(commandString string, caller *Player, lobby *Lobby) {
-	command := commands.ParseCommand(commandString)
-	if len(command) >= 1 {
-		switch strings.ToLower(command[0]) {
-		case "setmp":
-			commandSetMP(caller, lobby, command)
-		case "help":
-			//TODO
-		case "nick", "name", "username", "nickname", "playername", "alias":
-			commandNick(caller, lobby, command)
-		}
-	}
-}
-
-func commandNick(caller *Player, lobby *Lobby, args []string) {
-	if len(args) == 1 {
-		caller.Name = GeneratePlayerName()
-		WriteAsJSON(caller, JSEvent{Type: "reset-username"})
-		triggerPlayersUpdate(lobby)
-	} else {
-		//We join all arguments, since people won't sue quotes either way.
-		//The input is trimmed and sanitized.
-		newName := html.EscapeString(strings.TrimSpace(strings.Join(args[1:], " ")))
-		if len(newName) == 0 {
-			caller.Name = GeneratePlayerName()
-			WriteAsJSON(caller, JSEvent{Type: "reset-username"})
-		} else {
-			fmt.Printf("%s is now %s\n", caller.Name, newName)
-			//We don't want super-long names
-			if len(newName) > 30 {
-				newName = newName[:31]
-			}
-			caller.Name = newName
-			WriteAsJSON(caller, JSEvent{Type: "persist-username", Data: newName})
-		}
-		triggerPlayersUpdate(lobby)
-	}
-}
-
-func commandSetMP(caller *Player, lobby *Lobby, args []string) {
-	if caller == lobby.Owner {
-		if len(args) < 2 {
-			return
-		}
-
-		newMaxPlayersValue := strings.TrimSpace(args[1])
-		newMaxPlayersValueInt, err := strconv.ParseInt(newMaxPlayersValue, 10, 64)
-		if err == nil {
-			if int(newMaxPlayersValueInt) >= len(lobby.Players) && newMaxPlayersValueInt <= LobbySettingBounds.MaxMaxPlayers && newMaxPlayersValueInt >= LobbySettingBounds.MinMaxPlayers {
-				lobby.MaxPlayers = int(newMaxPlayersValueInt)
-
-				WritePublicSystemMessage(lobby, fmt.Sprintf("MaxPlayers value has been changed to %d", lobby.MaxPlayers))
-			} else {
-				if len(lobby.Players) > int(LobbySettingBounds.MinMaxPlayers) {
-					WriteAsJSON(caller, JSEvent{Type: "system-message", Data: fmt.Sprintf("MaxPlayers value should be between %d and %d.", len(lobby.Players), LobbySettingBounds.MaxMaxPlayers)})
-				} else {
-					WriteAsJSON(caller, JSEvent{Type: "system-message", Data: fmt.Sprintf("MaxPlayers value should be between %d and %d.", LobbySettingBounds.MinMaxPlayers, LobbySettingBounds.MaxMaxPlayers)})
-				}
-			}
-		} else {
-			WriteAsJSON(caller, JSEvent{Type: "system-message", Data: fmt.Sprintf("MaxPlayers value must be numeric.")})
-		}
-	} else {
-		WriteAsJSON(caller, JSEvent{Type: "system-message", Data: fmt.Sprintf("Only the lobby owner can change MaxPlayers setting.")})
-	}
+// AppendFill adds a fill direction to the current drawing. This exists in order
+// to prevent adding arbitrary elements to the drawing, as the backing array is
+// an empty interface type.
+func (lobby *Lobby) AppendFill(fill *FillEvent) {
+	lobby.CurrentDrawing = append(lobby.CurrentDrawing, fill)
 }
 
 func endRound(lobby *Lobby) {
@@ -424,7 +293,7 @@ func endRound(lobby *Lobby) {
 	//If the round ends and people still have guessing, that means the "Last" value
 	////for the next turn has to be "no score earned".
 	for _, otherPlayer := range lobby.Players {
-		if otherPlayer.State == Guessing {
+		if otherPlayer.State == PlayerStateGuessing {
 			otherPlayer.LastScore = 0
 		}
 	}
@@ -436,7 +305,7 @@ func endRound(lobby *Lobby) {
 
 func advanceLobby(lobby *Lobby) {
 	for _, otherPlayer := range lobby.Players {
-		otherPlayer.State = Guessing
+		otherPlayer.State = PlayerStateGuessing
 		otherPlayer.votedForKick = make(map[string]bool)
 	}
 
@@ -460,7 +329,7 @@ func advanceLobby(lobby *Lobby) {
 	}
 	selectNextDrawer(lobby)
 
-	lobby.Drawer.State = Drawing
+	lobby.Drawer.State = PlayerStateDrawing
 	lobby.WordChoice = GetRandomWords(lobby)
 
 	recalculateRanks(lobby)
@@ -476,7 +345,7 @@ func advanceLobby(lobby *Lobby) {
 		RoundEndTime: lobby.RoundEndTime,
 	}, lobby)
 
-	WriteAsJSON(lobby.Drawer, &JSEvent{Type: "your-turn", Data: lobby.WordChoice})
+	WriteAsJSON(lobby.Drawer, &GameEvent{Type: "your-turn", Data: lobby.WordChoice})
 }
 
 func selectNextDrawer(lobby *Lobby) {
@@ -557,134 +426,6 @@ func recalculateRanks(lobby *Lobby) {
 	}
 }
 
-func createWordHintFor(word string, showAll bool) []*WordHint {
-	wordHints := make([]*WordHint, 0, len(word))
-	for _, char := range word {
-		irrelevantChar := char == ' ' || char == '_' || char == '-'
-		if showAll {
-			wordHints = append(wordHints, &WordHint{
-				Character: char,
-				Underline: !irrelevantChar,
-			})
-		} else {
-			wordHints = append(wordHints, &WordHint{
-				Underline: !irrelevantChar,
-			})
-		}
-	}
-
-	return wordHints
-}
-
-var TriggerSimpleUpdateEvent func(eventType string, lobby *Lobby)
-var TriggerComplexUpdatePerPlayerEvent func(eventType string, data func(*Player) interface{}, lobby *Lobby)
-var TriggerComplexUpdateEvent func(eventType string, data interface{}, lobby *Lobby)
-var SendDataToConnectedPlayers func(sender *Player, lobby *Lobby, data interface{})
-var WriteAsJSON func(player *Player, object interface{}) error
-var WritePublicSystemMessage func(lobby *Lobby, text string)
-
-func triggerPlayersUpdate(lobby *Lobby) {
-	TriggerComplexUpdateEvent("update-players", lobby.Players, lobby)
-}
-
-func triggerCorrectGuessEvent(lobby *Lobby) {
-	TriggerSimpleUpdateEvent("correct-guess", lobby)
-}
-
-func triggerWordHintUpdate(lobby *Lobby) {
-	if lobby.CurrentWord == "" {
-		return
-	}
-
-	TriggerComplexUpdatePerPlayerEvent("update-wordhint", func(player *Player) interface{} {
-		return lobby.GetAvailableWordHints(player)
-	}, lobby)
-}
-
-type Rounds struct {
-	Round     int `json:"round"`
-	MaxRounds int `json:"maxRounds"`
-}
-
-// CreateLobby allows creating a lobby, optionally returning errors that
-// occured during creation.
-func CreateLobby(playerName, language string, drawingTime, rounds, maxPlayers, customWordChance, clientsPerIPLimit int, customWords []string, enableVotekick bool) (string, *Lobby, error) {
-	lobby := createLobby(drawingTime, rounds, maxPlayers, customWords, customWordChance, clientsPerIPLimit, enableVotekick)
-	player := createPlayer(playerName)
-
-	lobby.Players = append(lobby.Players, player)
-	lobby.Owner = player
-
-	// Read wordlist according to the chosen language
-	words, err := readWordList(language)
-	if err != nil {
-		//TODO Remove lobby, since we errored.
-		return "", nil, err
-	}
-
-	lobby.Words = words
-
-	return player.userSession, lobby, nil
-}
-
-// GeneratePlayerName creates a new playername. A so called petname. It consists
-// of an adverb, an adjective and a animal name. The result can generally be
-// trusted to be sane.
-func GeneratePlayerName() string {
-	adjective := strings.Title(petname.Adjective())
-	adverb := strings.Title(petname.Adverb())
-	name := strings.Title(petname.Name())
-	return adverb + adjective + name
-}
-
-// Message represents a message in the chatroom.
-type Message struct {
-	// Author is the player / thing that wrote the message
-	Author string `json:"author"`
-	// Content is the actual message text.
-	Content string `json:"content"`
-}
-
-// Ready represents the initial state that a user needs upon connection.
-// This includes all the necessary things for properly running a client
-// without receiving any more data.
-type Ready struct {
-	PlayerID string `json:"playerId"`
-	Drawing  bool   `json:"drawing"`
-
-	OwnerID        string        `json:"ownerId"`
-	Round          int           `json:"round"`
-	MaxRound       int           `json:"maxRounds"`
-	RoundEndTime   int64         `json:"roundEndTime"`
-	WordHints      []*WordHint   `json:"wordHints"`
-	Players        []*Player     `json:"players"`
-	CurrentDrawing []interface{} `json:"currentDrawing"`
-}
-
-func OnConnected(lobby *Lobby, player *Player) {
-	player.Connected = true
-	WriteAsJSON(player, JSEvent{Type: "ready", Data: &Ready{
-		PlayerID: player.ID,
-		Drawing:  player.State == Drawing,
-
-		OwnerID:        lobby.Owner.ID,
-		Round:          lobby.Round,
-		MaxRound:       lobby.MaxRounds,
-		RoundEndTime:   lobby.RoundEndTime,
-		WordHints:      lobby.GetAvailableWordHints(player),
-		Players:        lobby.Players,
-		CurrentDrawing: lobby.CurrentDrawing,
-	}})
-
-	//This state is reached when the player refreshes before having chosen a word.
-	if lobby.Drawer == player && lobby.CurrentWord == "" {
-		WriteAsJSON(lobby.Drawer, &JSEvent{Type: "your-turn", Data: lobby.WordChoice})
-	}
-
-	//TODO Only send to everyone except for the new player, since it's part of the ready event.
-	triggerPlayersUpdate(lobby)
-}
-
 func OnDisconnected(lobby *Lobby, player *Player) {
 	//We want to avoid calling the handler twice.
 	if player.ws == nil {
@@ -706,7 +447,7 @@ func (lobby *Lobby) GetAvailableWordHints(player *Player) []*WordHint {
 	//The draw simple gets every character as a word-hint. We basically abuse
 	//the hints for displaying the word, instead of having yet another GUI
 	//element that wastes space.
-	if player.State == Drawing || player.State == Standby {
+	if player.State == PlayerStateDrawing || player.State == PlayerStateStandby {
 		return lobby.WordHintsShown
 	} else {
 		return lobby.WordHints
